requirements.txt
typer[all]
rich
SQLAlchemy
alembic
bcrypt
python-dotenv
pytest
tabulate
